use ILGenerator() to compile script instead of intermediate C# file
generated code gets put in a ScriptObjCode object (analagous to DLL file)
then that gets put in a ScriptWrapper object for each instance

no DLL file is created.  also, as it stands now, script has to be
compiled every time for the first access to an assetID, as there is
no way for it to save the ScriptObjCode.  If ScriptObjCode could be
serialized to a file, then we wouldnt need to compile each time.

requires updated mono to run: mmrcont_load.c, writemltfile.c