change stand-alone script compiler to output abstract syntax tree instead of CIL-level code, so haxxors can't inject arbitrary CIL.
to build regular engine stuff, just go to Engine directory and
type a plain 'make' command without args.  it should create a
bunch of .cs files as usual, plus it should build a
../../../bin/xmrengshare.dll file that goes in same directory
as all the other OpenSim DLLs.

to build the stand-alone compiler, go to Compiler directory and
type a plain 'make' command without args.  it should create an
xmrengcomp.zip file that people can download, that contains the
xmrengcomp.exe compiler and some DLLs.  it also creates an
xmrengcomp_secret.exe file that can be used to dump out the
encrypted binaries created by xmrengcomp.exe (as well as do the
compiling stuff).