strip out properties
There is no easy or moderate way to account for non-trivial properties and
insert a CallLabel before the calls, as they can simply be read with the
evaluation stack in any state.  So maybe some step between the reducer and
the codegen can convert properties to get() and set() calls so the codegen
will know how to insert CallLabels in the code.