get comms stuff to work
The MethodInfos we get from GetScriptInvocation() have the exact same signature
as the methods in CommsTest.cs including the two UUID parameters tacked on the
front.  So I don't know why RegisterScriptInvocation() has to dynamically build
a delegate type, and if it weren't for all that nonsense we could call the
comms methods directly.

Also seems LookupModInvocation() could check for return type void and return
"modInvokeN".

Need to add 'AllowMODFunctions = true' to [XMREngine] in OpenSim.ini file.