base .xmrobj file on source text hash rather than script asset ID
...so it will use same in-memory object code for a given source
in case the script source file was copied around rather than
linked to.

if source file is completely empty, then it uses asset ID as
before to get the .xmrobj file.

source file can also contain just the hash string in which case
it is used directly as the .xmrobj filename.